LJH/requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Cython==3.0.10
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
Faker==24.3.0
fastapi==0.110.0
fastjsonschema==2.19.1
fasttext-wheel==0.9.2
filelock==3.13.1
fsspec==2024.3.0
greenlet==3.0.3
h11==0.14.0
hangul-utils==0.2
idna==3.6
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
networkx==3.2.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.3
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pybind11==2.12.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyMySQL==1.1.0
PyNaCl==1.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.1
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
threadpoolctl==3.3.0
tinycss2==1.2.1
torch==2.2.1
tornado==6.4
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.28.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
yarg==0.1.9
